% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{apply_gtm_per_variant}
\alias{apply_gtm_per_variant}
\title{Apply GTM to individual variants}
\usage{
apply_gtm_per_variant(counts, pr_beta = c(2000, 2000, 36, 12, 80, 1),
  pr_intv = rep(NA, 6), niter = 100, burnin = 10,
  two_sided = FALSE, independent = FALSE, cores = 1)
}
\arguments{
\item{counts}{list containing allele count data for each variant}

\item{pr_beta}{vector, parameters for beta priors}

\item{pr_intv}{determines if domain should be truncated, by default it is not}

\item{niter}{number of iterations}

\item{burnin}{number of burn-in steps}

\item{two_sided}{if TRUE, use two-sided beta priors}

\item{independent}{independent}
}
\value{
list, the GTM result for each variant
}
\description{
Apply GTM to each of a list of variants
}
