% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{construct_posteriors_frame}
\alias{construct_posteriors_frame}
\title{Construct posteriors data frame}
\usage{
construct_posteriors_frame(res_list)
}
\arguments{
\item{res_list}{list of GTM results}
}
\value{
data frame, posterior probabilities from GTM for each variant
}
\description{
convert GTM results to a data frame of variants and posterior
  probabilities
}
