% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASE_29Dec2014.R
\name{count.states}
\alias{count.states}
\title{count states}
\usage{
count.states(m, model.strong.ase = TRUE)
}
\arguments{
\item{m}{integer, the number of tissues}

\item{model.strong.ase}{logical, if TRUE then there are three groups, if FALSE then only two groups}
}
\description{
Counts how many heterogeneous states there are as a function of distance from homogeneous states.
}
\details{
MP 24-Feb-2014, 29-Dec-2014

Distance is the smallest number of changes that turns the state into one of the homogeneous states.
Thus maximum distance is m-ceiling(m/3) when model.strong.ase==TRUE, and floor(m/2) when model.strong.ase==FALSE
 minimum is 1 (for a heterogeneous state).
}
