% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extract_posteriors_row}
\alias{extract_posteriors_row}
\title{Extract Posteriors Row}
\usage{
extract_posteriors_row(single_variant_result)
}
\arguments{
\item{single_variant_result}{the GTM result for a single variant}
}
\value{
data frame, a row of the posteriors table
}
\description{
Extract a row of the posteriors table from GTM results
}
