% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASE_29Dec2014.R
\name{gtm.het}
\alias{gtm.het}
\title{Groupted Tissue Model Heterogeneity}
\usage{
gtm.het(y, prob, pr.beta, pr.intv, two.sided, independent, log.prior)
}
\description{
To check some of the most probable heterogeneity configurations based on individual probabilities
}
\details{
Taken these into account in marg likelihood calculation will improve accuracy of het probabilities
}
