% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASE_29Dec2014.R
\name{gtm.single}
\alias{gtm.single}
\title{Grouped Tissue Model Single}
\usage{
gtm.single(y, pr.beta, pr.intv, two.sided, model.strong.ase)
}
\value{
group probabilities for a single tissue
}
\description{
returns group probabilities for a single tissue
}
