% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASE_29Dec2014.R
\name{logmlk.3.truncated}
\alias{logmlk.3.truncated}
\title{truncated log marginal likelihood}
\usage{
logmlk.3.truncated(y, gr, pr.beta, pr.intv, two.sided, independent)
}
\value{
marginal loglikelihood
}
\description{
returns marginal loglikelihood
}
\details{
for input parameters see 'gtm'. allows for truncated priors and independence across tissues
}
