% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ase_29dec2014.R
\name{logprior.distance}
\alias{logprior.distance}
\title{Log Prior Distance}
\usage{
logprior.distance(m, p0 = 0.75, p.dist = NULL,
  model.strong.ase = TRUE)
}
\arguments{
\item{m}{the number of tissues}

\item{p0}{the joint prior probability of the 3 homogeneous states}

\item{p.dist}{either a vector of length 'max.dist' of total probabilities of each
set of states for distances 1,...,'max.dist'
Where max.dist==m-ceiling(m/3) if model.strong.ase==TRUE and max.dist==floor(m/2) if model.strong.ase==FALSE.
Interpreted as relative to each other so that after renormalisation and scaling sum to (1-p0).
OR, if p.dist == NULL,
  then p.dist will be set uniform =(1-p0)/max.dist over the distance.}
}
\value{
\describe{
  \item{log.prior}{
    is the log of prior probability of each STATE as a function of distance 1,...,max.dist
    results from dividing p.dist by the number of states in each distance category
  }
  \item{log.sum.prior.h0}{is the log of the sum of priors over all heterog states that have at least one 0}
  \item{log.sum.prior.h1}{is the log of the sum of priors over all heterog states that have no 0}
}
}
\description{
Counts the prior probability for each heterogeneous state as a function of its distance.
}
\details{
Distance is the smallest number of changes that turns the state into one of the homogeneous states
if model.strong.ase==TRUE then the maximum distance is m-ceiling(m/3),
if model.strong.ase==FALSE then the maximum distance is floor(m/2),
minimum distance is 1 for a heterogeneous configuration
(the three homogeneous configurations have distance 0)
}
